refactor(cli): rename common messages and parsers for all the commands under common name
