+++ packages/cli/src/commands/codegen.ts
@@ -50,8 +50,8 @@ export const codegen: Command = {
         `${intlMsg.commands_codegen_options_s()}`
       )
       .option(
-        `-c, --client-config <${intlMsg.commands_run_options_configPath()}>`,
-        `${intlMsg.commands_run_options_config()}`
+        `-c, --client-config <${intlMsg.commands_common_options_configPath()}>`,
+        `${intlMsg.commands_common_options_config()}`
       )
       .action(async (options) => {
         await run({
