+++ packages/cli/src/commands/run.ts
@@ -32,8 +32,8 @@ export const run: Command = {
         intlMsg.commands_run_options_workflowScript()
       )
       .option(
-        `-c, --client-config <${intlMsg.commands_run_options_configPath()}>`,
-        `${intlMsg.commands_run_options_config()}`
+        `-c, --client-config <${intlMsg.commands_common_options_configPath()}>`,
+        `${intlMsg.commands_common_options_config()}`
       )
       .option(
         `-v, --validate-script <${intlMsg.commands_run_options_validate()}>`,
