+++ packages/cli/src/lib/option-parsers/index.ts
@@ -4,3 +4,4 @@ export * from "./codegen";
 export * from "./wasm";
 export * from "./plugin";
 export * from "./run";
+export * from "./common";
